+++ src/pages/Dashboard/Payment/index.js
@@ -1,9 +1,37 @@
-import { useState } from 'react';
 import ChooseTicket from './ChooseTicket';
 import FinishPayment from './FinishPayment';
+import { useContext } from 'react';
+import { useState, useEffect } from 'react';
+import UserContext from '../../../contexts/UserContext';
+import { getUserTicket } from '../../../services/payment';
 
 export default function Payment() {
   const [finishPayment, setFinishPayment] = useState(false);
+  const [userTicket, setUserTicket] = useState(null);
+  const { userData } = useContext(UserContext);
+  const { token } = userData;
 
-  return <>{finishPayment ? <FinishPayment /> : <ChooseTicket setFinishPayment={setFinishPayment} />}</>;
+  useEffect(() => {
+    const fetchUserTicket = async() => {
+      try {
+        const ticket = await getUserTicket(token);
+        console.log('entrei');
+        setUserTicket(ticket.data);
+      } catch (err) {
+        console.log(err.message);
+      }
+    };
+
+    fetchUserTicket();
+  }, [finishPayment]);
+
+  return (
+    <>
+      {userTicket ? (
+        <FinishPayment userTicket={userTicket} />
+      ) : (
+        <ChooseTicket setFinishPayment={setFinishPayment} />
+      )}
+    </>
+  );
 }
